[fix-only] Ensure the fix for broken table content is on a separate line
We speculatively 'rescue' a broken HTML table with a fragment to close any interrupted tags, leaving HTML linting to remove the fragment if it isn't required.

At some point our code line that does the HTML linting started to snip the entire line out, removing the truncation notice too.

Putting the notice on a new line prevents this.